Fix/ fix type cpf in table User

+++ prisma/migrations/20230517192551_init/migration.sql
@@ -2,31 +2,37 @@
 CREATE TABLE "User" (
     "id" SERIAL NOT NULL,
     "name" TEXT NOT NULL,
-    "cpf" INTEGER NOT NULL,
+    "cpf" BIGINT NOT NULL,
     "email" TEXT NOT NULL,
     "password" TEXT NOT NULL,
+
     CONSTRAINT "User_pkey" PRIMARY KEY ("id")
 );
+
 -- CreateTable
 CREATE TABLE "Address" (
     "id" SERIAL NOT NULL,
     "street" TEXT NOT NULL,
-    "number" INTEGER NOT NULL,
+    "number" BIGINT NOT NULL,
     "district" TEXT NOT NULL,
     "city" TEXT NOT NULL,
     "state" TEXT NOT NULL,
     "userId" INTEGER NOT NULL,
     "createdAt" TIMESTAMP(3) NOT NULL DEFAULT CURRENT_TIMESTAMP,
     "updatedAt" TIMESTAMP(3) NOT NULL,
+
     CONSTRAINT "Address_pkey" PRIMARY KEY ("id")
 );
+
 -- CreateTable
 CREATE TABLE "Category" (
     "id" SERIAL NOT NULL,
     "name" TEXT NOT NULL,
     "description" TEXT NOT NULL,
+
     CONSTRAINT "Category_pkey" PRIMARY KEY ("id")
 );
+
 -- CreateTable
 CREATE TABLE "Product" (
     "id" SERIAL NOT NULL,
@@ -36,8 +42,10 @@ CREATE TABLE "Product" (
     "quantity" INTEGER NOT NULL,
     "image" TEXT NOT NULL,
     "categoryId" INTEGER NOT NULL,
+
     CONSTRAINT "Product_pkey" PRIMARY KEY ("id")
 );
+
 -- CreateTable
 CREATE TABLE "Order" (
     "id" SERIAL NOT NULL,
@@ -45,8 +53,10 @@ CREATE TABLE "Order" (
     "total" DOUBLE PRECISION NOT NULL,
     "createdAt" TIMESTAMP(3) NOT NULL DEFAULT CURRENT_TIMESTAMP,
     "updatedAt" TIMESTAMP(3) NOT NULL,
+
     CONSTRAINT "Order_pkey" PRIMARY KEY ("id")
 );
+
 -- CreateTable
 CREATE TABLE "Admin" (
     "id" SERIAL NOT NULL,
@@ -54,24 +64,30 @@ CREATE TABLE "Admin" (
     "email" TEXT NOT NULL,
     "password" TEXT NOT NULL,
     "role" TEXT NOT NULL,
+
     CONSTRAINT "Admin_pkey" PRIMARY KEY ("id")
 );
+
 -- CreateIndex
 CREATE UNIQUE INDEX "User_id_key" ON "User"("id");
+
 -- CreateIndex
 CREATE UNIQUE INDEX "User_cpf_key" ON "User"("cpf");
+
 -- CreateIndex
 CREATE UNIQUE INDEX "User_email_key" ON "User"("email");
+
 -- CreateIndex
 CREATE UNIQUE INDEX "User_password_key" ON "User"("password");
+
 -- CreateIndex
 CREATE UNIQUE INDEX "Admin_email_key" ON "Admin"("email");
+
 -- AddForeignKey
-ALTER TABLE "Address"
-ADD CONSTRAINT "Address_userId_fkey" FOREIGN KEY ("userId") REFERENCES "User"("id") ON DELETE RESTRICT ON UPDATE CASCADE;
+ALTER TABLE "Address" ADD CONSTRAINT "Address_userId_fkey" FOREIGN KEY ("userId") REFERENCES "User"("id") ON DELETE RESTRICT ON UPDATE CASCADE;
+
 -- AddForeignKey
-ALTER TABLE "Product"
-ADD CONSTRAINT "Product_categoryId_fkey" FOREIGN KEY ("categoryId") REFERENCES "Category"("id") ON DELETE RESTRICT ON UPDATE CASCADE;
+ALTER TABLE "Product" ADD CONSTRAINT "Product_categoryId_fkey" FOREIGN KEY ("categoryId") REFERENCES "Category"("id") ON DELETE RESTRICT ON UPDATE CASCADE;
+
 -- AddForeignKey
-ALTER TABLE "Order"
-ADD CONSTRAINT "Order_userId_fkey" FOREIGN KEY ("userId") REFERENCES "User"("id") ON DELETE RESTRICT ON UPDATE CASCADE;
+ALTER TABLE "Order" ADD CONSTRAINT "Order_userId_fkey" FOREIGN KEY ("userId") REFERENCES "User"("id") ON DELETE RESTRICT ON UPDATE CASCADE;
